Rename graphql-loader -> products-loader
